backend/requirements.txt
# Core Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0

# Database
sqlalchemy[asyncio]>=2.0.37,<3.0.0
asyncpg>=0.30.0,<1.0.0
alembic>=1.14.0,<2.0.0

# Redis
redis[hiredis]>=5.2.0,<6.0.0
arq>=0.26.1,<1.0.0

# AI
litellm>=1.60.0,<2.0.0

# Auth
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<4.1.0
email-validator>=2.1.0,<3.0.0

# Cloud Providers
boto3>=1.34.0
asyncssh>=2.14.0
google-cloud-logging>=3.10.0
google-cloud-compute>=1.15.0
google-auth>=2.25.0

# HTTP Client
httpx>=0.28.0,<1.0.0

# Observability
structlog>=24.4.0,<26.0.0
prometheus-client>=0.21.0,<1.0.0

# SSE
sse-starlette>=2.2.0,<3.0.0

# Utilities
python-multipart>=0.0.20,<1.0.0
pyyaml>=6.0

# Testing
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.25.0,<1.0.0
httpx>=0.28.0,<1.0.0
